refactor of App.tsx
